requirements.txt
flickrapi==2.4.0
  requests==2.31.0
    certifi==2024.2.2
    charset-normalizer==3.3.2
    idna==3.7
    urllib3==2.2.1
  requests-oauthlib==2.0.0
    oauthlib==3.2.2
    requests==2.31.0
      certifi==2024.2.2
      charset-normalizer==3.3.2
      idna==3.7
      urllib3==2.2.1
  requests-toolbelt==1.0.0
    requests==2.31.0
      certifi==2024.2.2
      charset-normalizer==3.3.2
      idna==3.7
      urllib3==2.2.1
  six==1.16.0
icecream==2.1.3
  asttokens==2.4.1
    six==1.16.0
  colorama==0.4.6
  executing==2.0.1
  Pygments==2.17.2
openpyxl==3.1.2
  et-xmlfile==1.1.0
pandas==2.2.2
  numpy==1.26.4
  python-dateutil==2.9.0.post0
    six==1.16.0
  pytz==2024.1
  tzdata==2024.1
pipdeptree==2.20.0
  packaging==24.0
  pip==24.0
PyQt5==5.15.10
  PyQt5-Qt5==5.15.13
  PyQt5-sip==12.13.0
python-dotenv==1.0.1
